FreeSky/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.4
django-ckeditor==6.4.0
django-js-asset==2.0.0
djangorestframework==3.13.1
environs==9.5.0
idna==3.3
marshmallow==3.15.0
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
psycopg2==2.9.3
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.9
yaschedule==0.0.4.0
